backend/requirements.txt
Django==4.2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-ratelimit==4.1.0
django-storages==1.14.2
channels==4.0.0
channels-redis==4.2.0
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
cloudinary==1.39.0
reportlab==4.1.0
WeasyPrint==62.3
twilio==8.10.3
sendgrid==6.11.0
gunicorn==21.2.0
python-dotenv==1.0.1
razorpay==1.4.1
pytest-django==4.7.0
factory-boy==3.3.0
